requirements.txt
opencv-python~=4.6.0.66
numpy~=1.18.5
py2exe~=0.11.0.1
matplotlib~=3.3.4
Pillow~=8.4.0
keras~=2.10.0
scikit-image~=0.17.2
scikit-learn~=0.24.2
tqdm~=4.64.1